crbtree: add traversal helpers
Add three new static-inline helpers to the rbtree API. They implement
tree-traversal, based on a comparison-function. They differ in the the
pointer type they return.

These helpers are in no way mandatory, but rather examples how to code
insertion and lookup functions. An elaborate test is added for the new
code.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>