c-rbtree: implement post-order traversal
Add 6 new helpers to perform post-order tree traversal. They return the
first, last, next, and prev elements for post-order tree traversals. We
only implement left-to-right post-order traversals, as usually there is
no need for a specific direction. If right-to-left is needed, you better
open-code it yourself.

Note that an inverse left-to-right post-order traversal can be
interpreted as a right-to-left pre-order traversal. So this set of
helpers is enough to implement almost all common kinds of tree traversals
(ignoring the direction, and lacking support for level-order, which is
pretty uncommon for balanced binary trees, anyway).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>