c-rbtree: support lockless lookups
This adds support for lockless tree lookups. So this is tricky. If you
have a shared tree between multiple different CPUs, you usually need a
reader/writer lock to protect the tree integrity. There is a faster
option, though: If you protect your tree nodes via RCU, you can avoid the
rw-lock and instead use a lockless RCU lookup. The writer side still needs
mutual exclusion, but the reader side can run without taking any lock.

But this is not the whole story. A tree rotation is never atomic (requires
multiple pointers to be changed). Therefore, a concurrent reader might get
an inconsistent view of a tree during insertions or removals, and there is
no way we can avoid that. However, we can modify the tree helpers to
guarantee that a reader never sees partial stores nor hits cycles in the
tree. Thus, a read can at least get a valid, well-defined result when
running unlocked, but needs to be aware that it possibly skipped entire
sub-trees during the traversal. So it might get false-negatives, but never
false-positives. If false-negatives cannot be accepted, a simple seqlock
(i.e., retry loop) can be employed to simply retry the lookup if it really
raced a modification.

This commit changes the INSERTION and REMOVAL functions of the RBTree to
give the following guarantees:

        * all n->{left,right} pointers are stored atomically and are
          always valid node pointers

        * there is never any cycle in an rb-tree regarding the
          n->{left,right} pointers

Note that parent pointers might still produce cycles! However, a normal
lookup does not need the parent pointers. We only guarantee that the
left/right pointers to not contain cycles. If you need more fancy lookups,
this will not work for you.

Since the cycle-detection is non-trivial to review, this commit includes a
test that verifies on random trees that they never contain any cycles. It
uses ptrace to step through the insert/remove helpers instruction by
instruction, and after each instruction verify the tree for correctness.

This commit does not introduce the required memory barriers for stores and
reads. This is left to the caller for now, but might be added in later
commits.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>