tests: add tsearch(3p) comparison
The glibc tsearch(3p) API uses RB-Trees, so lets add a test to compare it
against c-rbtree.

Note that the tsearch(3p) API in itself is very restricted and useless
for most scenarios (e.g., twalk() does not have any caller context). So
this is mostly a purely academic comparison.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>