c-rbtree: turn color into generic flag
Rather than having a single color-marker, we now have a generic flag
field on each node. This allows us to extend the flags.

This is a rather dumb search-and-replace commit. No additional logic is
applied to the unused flag fields. When adding flags, a full audit
needs to be performed, to ensure they are retained correctly in all
cases.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>