c-rbtree: remove bogus comment
The '!n' case does not imply 'go to right'. On the contrary, '!n' means
we came from an empty child. Hence, the comparison done is still
correct. In case both children are NULL, it does not matter which path
we take. But also, we would never be called if both children are NULL.
So yeah, remove the bogus comment.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>