c-rbtree: embedd tree-root in root-node
Many intrusive collections allow node removal without access to the
surrounding context. That is, adding a node to a context obviously
requires the context. However, removing it again should work without
access to the context.

Our c-list API does this nicely. Linking is done via c_list_link()
(allowing with or without context access), but unlinking is easy via
c_list_unlink(CList *node). No separate list pointer required, just the
node.

In case of c-rbtree we always wanted the same. A simple way to call
c_rbnode_unlink(CRBNode *node) to remove the node. However, the issue
always was that removing the root-node requires updating the tree-root.
But the nodes themselves do not embed the tree-root.

There are 2 obvious ways to fix that:

  1) Embed a pointer to the tree-root in the root-node. Thus, modifying
     the root-node grants access to the tree-root.

  2) Make the tree-root part of the tree.

The second solution is the more elegant one, but with the downside that
tree-traversals need to be aware that the tree-root is not a valid
node. The c-list helpers do something similar to 2).
Furthermore, 2) also has the downside that the tree-root might no
longer actually be the root-node. This could be fixed by rotating with
a fixed tree-root. However, this easily becomes confusing.
Nevertheless, this solution would still provide O(log(n)) access.

The first solution is less elegant, but with the upside of keeping the
API clean. Hence, this is what is implemented here:

    The root-node of any tree now uses the 'parent' pointer to store a
    pointer to the tree-root. We introduce an additional flag which
    marks a root-node as such. This way, whenever we modify it, we know
    whether it is the root-node, and additionally get a pointer to the
    tree-root.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>