c-rbtree: rename c_rbnode_unlink{,_init}
Make c_rbnode_unlink() work like c_rbnode_unlink_init() used to, and rename the
old c_rbnode_unlink() to c_rbnode_unlink_stale(). This is meant to encourage the
use of the former, and for people to only use the stale version in the extremely
rare case that it affects performance.

Signed-off-by: Tom Gundersen <teg@jklm.no>