fix build on m68k/sh4 archs and require minimal struct aligment to only 4 bytes
Building on m68k and sh4 architectures fails with static assertions:

    shared/c-rbtree/src/c-rbtree.c:39:1: error: static assertion failed: "Invalid RBNode alignment"
       39 | static_assert(alignof(CRBNode) <= alignof(max_align_t), "Invalid RBNode alignment");
          | ^~~~~~~~~~~~~
    shared/c-rbtree/src/c-rbtree.c:41:1: error: static assertion failed: "Invalid RBTree alignment"
       41 | static_assert(alignof(CRBTree) <= alignof(max_align_t), "Invalid RBTree alignment");
          | ^~~~~~~~~~~~~
    make[3]: *** [Makefile:13261: shared/c-rbtree/src/libcrbtree_la-c-rbtree.lo] Error 1
    make[3]: *** Waiting for unfinished jobs....

I presume that is, because on those architectures "max_align_t" is only aligned to 4 bytes.

The previous alignment to 8 bytes was to reserve 4 flags, and ensure that all valid
pointers have these lower bits unset. Since one flag was actually unused, we only
need to reserve 3 bits and can accept an alignment to 4 bytes only.

Also, while at it don't use "unsigned long" but uintptr_t type. Since this is supposed to
be a C11 library, we can rely on <stdint.h> to provide that type for us (on any architecture
we care about).

This is an ABI break, but I think that is acceptable, since c-rbtree is
usually statically linked.